+++ src/cpp11.cpp
@@ -713,6 +713,13 @@ extern "C" SEXP _cheapr_cpp_tabulate(SEXP x, SEXP n_bins) {
     return cpp11::as_sexp(cpp_tabulate(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x), cpp11::as_cpp<cpp11::decay_t<uint32_t>>(n_bins)));
   END_CPP11
 }
+// utils.cpp
+SEXP cpp_is_whole_number(SEXP x, double tol_, bool na_rm_);
+extern "C" SEXP _cheapr_cpp_is_whole_number(SEXP x, SEXP tol_, SEXP na_rm_) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(cpp_is_whole_number(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x), cpp11::as_cpp<cpp11::decay_t<double>>(tol_), cpp11::as_cpp<cpp11::decay_t<bool>>(na_rm_)));
+  END_CPP11
+}
 // which.cpp
 SEXP cpp_which_(SEXP x, bool invert);
 extern "C" SEXP _cheapr_cpp_which_(SEXP x, SEXP invert) {
@@ -786,6 +793,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_cheapr_cpp_is_na",                   (DL_FUNC) &_cheapr_cpp_is_na,                   1},
     {"_cheapr_cpp_is_simple_atomic_vec",    (DL_FUNC) &_cheapr_cpp_is_simple_atomic_vec,    1},
     {"_cheapr_cpp_is_simple_vec",           (DL_FUNC) &_cheapr_cpp_is_simple_vec,           1},
+    {"_cheapr_cpp_is_whole_number",         (DL_FUNC) &_cheapr_cpp_is_whole_number,         3},
     {"_cheapr_cpp_lag",                     (DL_FUNC) &_cheapr_cpp_lag,                     5},
     {"_cheapr_cpp_lag2",                    (DL_FUNC) &_cheapr_cpp_lag2,                    6},
     {"_cheapr_cpp_lag_sequence",            (DL_FUNC) &_cheapr_cpp_lag_sequence,            3},